+++ tests/testthat/test-verb-distinct.R
@@ -21,6 +21,7 @@ test_that("correctly inlines across all verbs", {
   # two table verbs
   lf2 <- lazy_frame(x = 1)
   expect_selects(lf |> left_join(lf2, by = "x") |> distinct(), 1)
+  expect_selects(lf |> right_join(lf2, by = "x") |> distinct(), 2)
   expect_selects(lf |> semi_join(lf2, by = "x") |> distinct(), 3)
   expect_selects(lf |> union(lf2) |> distinct(), 3)
 })