+++ tests/testthat/test-verb-select.R
@@ -38,6 +38,7 @@ test_that("correctly inlines across all verbs", {
   # two table verbs
   lf2 <- lazy_frame(x = 1)
   expect_selects(lf |> left_join(lf2, by = "x") |> select(x), 1)
+  expect_selects(lf |> right_join(lf2, by = "x") |> select(x), 1)
   expect_selects(lf |> semi_join(lf2, by = "x") |> select(x), 2)
   expect_selects(lf |> union(lf2) |> select(x), 3)
 })